moved news page html to component, created baic layout for news rss
